+++ ServerSideItems/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.0.1
+
+- Update R2API dependencies
+
 # 1.0.0
 
 - Initial Release